Bug 572833  protect the product with common-use license
Extract EclispePassageUI interactions for reuse
