+++ bundles/org.eclipse.passage.lic.jface/src/org/eclipse/passage/lic/internal/jface/actions/LicensedAction.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2020 ArSysOp
+ * Copyright (c) 2018, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -12,19 +12,10 @@
  *******************************************************************************/
 package org.eclipse.passage.lic.internal.jface.actions;
 
-import java.util.Optional;
-
 import org.eclipse.jface.action.Action;
-import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
-import org.eclipse.passage.lic.internal.api.access.GrantLockAttempt;
-import org.eclipse.passage.lic.internal.equinox.EquinoxPassage;
-import org.eclipse.passage.lic.internal.jface.EquinoxPassageUI;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Event;
 
-/**
- * @since 1.1
- */
 public abstract class LicensedAction extends Action {
 
 	protected abstract void doAction();
@@ -40,23 +31,7 @@ public abstract class LicensedAction extends Action {
 	}
 
 	private void runEverywhere(Display display) {
-		Optional<ServiceInvocationResult<GrantLockAttempt>> response = Optional.empty();
-		try {
-			response = Optional.of(new EquinoxPassageUI(display::getActiveShell).acquireLicense(getId()));
-			if (grantAcquired(response)) {
-				doAction();
-			}
-		} finally {
-			response.flatMap(ServiceInvocationResult::data)//
-					.ifPresent(lock -> new EquinoxPassage().releaseLicense(lock));
-		}
-	}
-
-	private boolean grantAcquired(Optional<ServiceInvocationResult<GrantLockAttempt>> response) {
-		return response//
-				.flatMap(ServiceInvocationResult::data)//
-				.map(GrantLockAttempt::successful)//
-				.orElse(false);
+		new LicensedRunnable(display::getActiveShell, getId(), this::doAction).run();
 	}
 
 }
